added support for config versioning
