+++ manager/config_parser/__init__.py
@@ -1,0 +1,1 @@
+from .config_parser import *
